# Conflicts:
#	src/main/resources/application.properties

+++ src/main/java/com/example/class_manager/repositories/TeacherRepository.java
@@ -3,5 +3,5 @@ package com.example.class_manager.repositories;
 import com.example.class_manager.models.Teacher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeacherRepository extends JpaRepository<Teacher, Long> {
+public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
 }
